+++ app/src/main/java/cn/atsoft/dasheng/crm/model/params/PaymentParam.java
@@ -74,7 +74,7 @@ public class PaymentParam implements Serializable, BaseValidatingParam {
      * 付款方式
      */
     @ApiModelProperty("付款方式")
-    private Integer payPlan;
+    private Long payPlan;
 
     /**
      * 创建者
